feat: merge certificates into SPA, redesign tiering, and switch to SVG den images
- Merge standalone certificate bundle into main React SPA (CertificateView)
  using React Router; delete certificate.html and certificate.tsx
- Add tiered certificates: podium (1-3), top5 (4-5), top10 (6-10),
  den champion, den top3, and achievement with placement subtitles
- Extract certificate stats logic into testable certificate-stats.ts
  with buildCertificateStats() and computeRacerStats()
- Replace PNG den images with SVGs for resolution-independent print quality
- Add 3-column grid layout for stats row to center den image precisely
- Show "Xth place of Y racers" subtitle on all non-podium tiers
- Achievement tier shows Races stat as balancing filler
- Ordinal suffixes render at 0.65em with reduced opacity
- Bump dialog text sizes globally (title xl bold, description base)
- Add standings ranking research notes and sources docs
