+++ docs/sources.md
@@ -1,0 +1,28 @@
+# Sources & References
+
+Useful references organized by topic, collected during development.
+
+## Pinewood Derby Ranking & Scoring
+
+- [Stan Pope - Conducting Pinewood Derbies](https://stanpope.net/pwraces.html) — Authoritative guide covering scheduling, scoring, and race management
+- [GrandPrix Race Manager (GPRM) - Standings FAQ](https://grandprix-software-central.com/index.php/faqs/gprm/standings) — Tiebreaker rules and scoring options
+- [GrandPrix Race Manager](https://www.grandprix-software-central.com/gprm/) — Most widely-used commercial derby software
+- [DerbyNet](https://derbynet.org/) ([GitHub](https://github.com/jeffpiazza/derbynet)) — Popular open-source alternative, uses "Partial Perfect-N" scheduling
+- [Maximum Velocity - Race Methods](https://www.maximum-velocity.com/pinewood-derby-race-methods/) — Elimination race method comparison
+
+## Derby Community Discussions
+
+- [Derby Talk - Points vs. Times Scoring](https://www.derbytalk.com/viewtopic.php?t=2510)
+- [Derby Talk - Tiebreaker Scenario](https://derbytalk.com/viewtopic.php?t=1882)
+- [Derby Talk - GPRM Tie Breaker Rules](https://www.derbytalk.com/viewtopic.php?t=8201)
+- [Scouting Magazine - Scheduling Strategies](https://scoutingmagazine.org/2015/11/scheduling-strategies-transform-packs-next-pinewood-derby/)
+
+## Motorsport Scoring Systems (for reference)
+
+- [RaceMate - F1 Tie Breakers Explained](https://racemate.io/blog/f1-tie-breakers-explained/) — F1 countback system (most wins, then most 2nds, etc.)
+- [Wikipedia - Motorsports Points Scoring Systems](https://en.wikipedia.org/wiki/List_of_motorsports_points_scoring_systems)
+- [Slot Car Club UK - Ranking Info](https://www.slotcarclub.co.uk/ranking-info/) — Points-per-event slot car rankings
+
+## Track Hardware
+
+- See `docs/electronics/` for timer integration docs and serial bus specs
