+++ src/frontend/assets.d.ts
@@ -1,0 +1,4 @@
+declare module '*.svg' {
+  const src: string;
+  export default src;
+}
