--- src/frontend/certificate.html
@@ -1,13 +1,0 @@
-<!doctype html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <title>Certificate - Derby Race Manager</title>
-        <link rel="stylesheet" href="./styles/styles.css" />
-    </head>
-    <body>
-        <div id="app"></div>
-        <script type="module" src="./certificate.tsx"></script>
-    </body>
-</html>
