+++ src/frontend/lib/den-utils.ts
@@ -1,15 +1,15 @@
-import lionImg    from '../assets/dens/lion-rank-normalized.png';
-import tigersImg  from '../assets/dens/tigers-rank-normalized.png';
-import wolvesImg  from '../assets/dens/wolves-rank-normalized.png';
-import bearsImg   from '../assets/dens/bears-rank-normalized.png';
-import webelosImg from '../assets/dens/webelos-rank-normalized.png';
-import aolImg     from '../assets/dens/aol-rank-normalized.png';
+import lionImg    from '../assets/dens/lion-rank.svg';
+import tigerImg   from '../assets/dens/tiger-rank.svg';
+import wolfImg    from '../assets/dens/wolf-rank.svg';
+import bearImg    from '../assets/dens/bear-rank.svg';
+import webelosImg from '../assets/dens/webelos-rank.svg';
+import aolImg     from '../assets/dens/aol-rank.svg';
 
 export const DEN_IMAGES: Record<string, string> = {
   'Lions':   lionImg,
-  'Tigers':  tigersImg,
-  'Wolves':  wolvesImg,
-  'Bears':   bearsImg,
+  'Tigers':  tigerImg,
+  'Wolves':  wolfImg,
+  'Bears':   bearImg,
   'Webelos': webelosImg,
   'AOLs':    aolImg,
 };
